LayoutTests/platform/wpe/imported/w3c/web-platform-tests/html/semantics/embedded-content/bfcache/embedded-not-found-expected.txt

CONSOLE MESSAGE: QuotaExceededError: The quota has been exceeded.

Harness Error (TIMEOUT), message = null

TIMEOUT Page with <embed type=image/png src=/404.png> Test timed out
NOTRUN Page with <object type=image/png data=/404.png>
